Rename stencil display-wrapper
https://coveord.atlassian.net/browse/KIT-4197
