+++ packages/atomic/src/components/search/result-lists/atomic-result-list/atomic-result-list.tsx
@@ -28,10 +28,10 @@ import {
   DisplayTable,
   DisplayTableRow,
 } from '../../../common/item-list/display-table';
-import {DisplayWrapper} from '../../../common/item-list/display-wrapper';
 import {ItemDisplayGuard} from '../../../common/item-list/item-display-guard';
 import {ItemListGuard} from '../../../common/item-list/item-list-guard';
 import {ItemTemplateProvider} from '../../../common/item-list/item-template-provider';
+import {DisplayWrapper} from '../../../common/item-list/stencil-display-wrapper';
 import {
   ItemListCommon,
   ItemRenderingFunction,
